Feat: can view md in proposals
